Sprint2

+++ src/main/java/Projeto/DAC/service/CarroService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import Projeto.DAC.model.Carro;
+import Projeto.DAC.model.Usuario;
 import Projeto.DAC.repository.CarroRepository;
 
 @Service
@@ -17,7 +18,19 @@ public class CarroService {
 	@Autowired
 	private CarroRepository carroRepository;
 	
-	public Carro salvar(Carro carro) {
+	@Autowired
+	public UsuarioService usuarioService;
+	
+	public Carro salvar(Carro carro, Long Id) {
+		
+		Usuario usuario = usuarioService.listarPorId(Id);
+		
+		String loginType = (usuario != null && usuario.getMatricula() != null) ? "ADMIN" : "USER";
+		if(loginType.equals("USER")) 
+		{
+			throw new IllegalStateException("Usuários comum não têm permissão para inserir carros.");
+		}
+		
 		return carroRepository.save(carro);
 	}
 	
@@ -33,14 +46,31 @@ public class CarroService {
 		return carroRepository.findById(id).orElseThrow( () -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Carro não encontrado") );
 	}
 	
-	public void excluir(Long id) {
+	public void excluir(Long id, Long usuarioId) {
+		
+		Usuario usuario = usuarioService.listarPorId(usuarioId);
+		
+		String loginType = (usuario != null && usuario.getMatricula() != null) ? "ADMIN" : "USER";
+		if(loginType.equals("USER")) 
+		{
+			throw new IllegalStateException("Usuários comum não têm permissão para inserir carros.");
+		}
+		
 		carroRepository.findById(id).map(carro -> {
 			carroRepository.delete(carro);
 	            return Void.TYPE;
 	        }).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Carro não encontrado"));
 	}
 	
-	public void editar(Long id, Carro carroAtualizado) {
+	public void editar(Long id, Carro carroAtualizado, Long usuarioId) {
+		
+		Usuario usuario = usuarioService.listarPorId(usuarioId);
+		
+		String loginType = (usuario != null && usuario.getMatricula() != null) ? "ADMIN" : "USER";
+		if(loginType.equals("USER")) 
+		{
+			throw new IllegalStateException("Usuários comum não têm permissão para inserir carros.");
+		}
 		carroRepository.findById(id)
         .map( carro -> {
         	carro.setModelo(carroAtualizado.getModelo());
